+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2019-03-17
+### Added
+- MSGraphFunctions now supports non-interactive authentication to Microsoft Graph using Delegated permissions
+
+### Changed
+- BREAKING: Connect-Graph now requires a $Credential parameter of type PSCredential
+
 ## [1.0.0] - 2019-03-15
 ### Added
 - PowerShell module initial release